backend/requirements.txt
asgiref==3.7.2
Django==3.2.21
django-rest-framework==0.1.0
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.7.1
